requirements.txt
appdirs==1.4.3
atomicwrites==1.2.1
attrs==18.2.0
black==18.9b0
certifi==2018.10.15
chardet==3.0.4
Click==7.0
idna==2.7
more-itertools==4.3.0
multidict==4.4.2
pluggy==0.8.0
py==1.7.0
pytest==3.10.0
PyYAML==3.13
requests==2.20.0
six==1.11.0
toml==0.10.0
urllib3==1.24.1
vcrpy==2.0.1
wrapt==1.10.11
yarl==1.2.6
